+++ lib/data/fuzzy_options.dart
@@ -9,7 +9,7 @@ class WeightedKey<T> {
     @required this.name,
     @required this.getter,
     @required this.weight,
-  }) : assert(weight >= 0 && weight <= 1);
+  }) : assert(weight > 0, 'Weight should be positive and non-zero');
 
   /// Name of this getter
   final String name;
@@ -17,7 +17,9 @@ class WeightedKey<T> {
   /// Getter to a specifc string inside item
   final String Function(T obj) getter;
 
-  /// Weight of this getter
+  /// Weight of this getter. When passing a list of WeightedKey to FuzzyOptions,
+  /// the weight can be any positive number; FuzzyOptions normalizes it on
+  /// construction.
   final double weight;
 }
 
@@ -28,7 +30,9 @@ int _defaultSortFn<T>(Result<T> a, Result<T> b) => a.score.compareTo(b.score);
 
 /// Options for performing a fuzzy search
 class FuzzyOptions<T> {
-  /// Instantiate an options object
+  /// Instantiate an options object.
+  /// The `keys` list requires a positive number (they'll be normalized upon
+  /// instantiation). If any weight is not positive, throws an ArgumentError.
   FuzzyOptions({
     this.location = 0,
     this.distance = 100,
@@ -36,9 +40,10 @@ class FuzzyOptions<T> {
     this.maxPatternLength = 32,
     this.isCaseSensitive = false,
     Pattern tokenSeparator,
+    this.minTokenCharLength = 1,
     this.findAllMatches = false,
     this.minMatchCharLength = 1,
-    this.keys = const [],
+    List<WeightedKey<T>> keys = const [],
     this.shouldSort = true,
     SorterFn<T> sortFn,
     this.tokenize = false,
@@ -47,6 +52,7 @@ class FuzzyOptions<T> {
     this.shouldNormalize = false,
   })  : tokenSeparator =
             tokenSeparator ?? RegExp(r' +', caseSensitive: isCaseSensitive),
+        keys = _normalizeWeights(keys),
         sortFn = sortFn ?? _defaultSortFn;
 
   /// Approximately where in the text is the pattern expected to be found?
@@ -71,6 +77,9 @@ class FuzzyOptions<T> {
 
   /// Regex used to separate words when searching. Only applicable when `tokenize` is `true`.
   final Pattern tokenSeparator;
+
+  /// Ignore tokens with length smaller than this. Only applicable when `tokenize` is `true`.
+  final int minTokenCharLength;
 
   /// When true, the algorithm continues searching to the end of the input even if a perfect
   /// match is found before the end of the same input.
@@ -112,6 +121,7 @@ class FuzzyOptions<T> {
         maxPatternLength: options?.maxPatternLength ?? maxPatternLength,
         isCaseSensitive: options?.isCaseSensitive ?? isCaseSensitive,
         tokenSeparator: options?.tokenSeparator ?? tokenSeparator,
+        minTokenCharLength: options?.minTokenCharLength ?? minTokenCharLength,
         findAllMatches: options?.findAllMatches ?? findAllMatches,
         minMatchCharLength: options?.minMatchCharLength ?? minMatchCharLength,
         keys: options?.keys ?? keys,
@@ -122,4 +132,22 @@ class FuzzyOptions<T> {
         verbose: options?.verbose ?? verbose,
         shouldNormalize: options?.shouldNormalize ?? shouldNormalize,
       );
+
+  static List<WeightedKey<T>> _normalizeWeights<T>(List<WeightedKey<T>> keys) {
+    if (keys.isEmpty) {
+      return [];
+    }
+
+    var weightSum = keys
+        .map((key) => key.weight)
+        .fold<double>(0, (previousValue, element) => previousValue + element);
+
+    return keys
+        .map((key) => WeightedKey<T>(
+              name: key.name,
+              getter: key.getter,
+              weight: key.weight / weightSum,
+            ))
+        .toList();
+  }
 }